dev debug 0.02

+++ frontend/controllers/ChepuhaBotController.php
@@ -342,11 +342,13 @@ class ChepuhaBotController extends \yii\web\Controller
 //          /end
             if (trim(strtolower($message['text'])) == '/end') {
                 $session = ChBotSession::find()->where(['user_id'=>$message['from']['id']])->one();
-                
+
                 if ($session != null) {
                     $vars = $session->vars;
-                    foreach ($vars as  $var) {
-                        $var->delete();
+                    if ($vars != null) {
+                        foreach ($vars as  $var) {
+                            $var->delete();
+                        }
                     }
                     $use = BotUse::find()->where(['id'=>intval($session['description'])])->one();
                     $session->delete();
